feat(Middleware): -remove try-catch in public async Task<IActionResult> GetById(Guid id)
