+++ Api/Middlewares/ExceptionHandlingMiddleware.cs
@@ -1,0 +1,45 @@
+﻿using Microsoft.Identity.Client;
+using System.Diagnostics.Eventing.Reader;
+using Microsoft.AspNetCore.Http;
+using Microsoft.Extensions.Logging;
+using System.Net;
+using System.Threading.Tasks;
+using System.Text.Json;
+
+namespace MyCompany.Test.Api.Middlewares
+{
+    public class ExceptionHandlingMiddleware
+    {
+        private readonly RequestDelegate _next;
+        private readonly ILogger<ExceptionHandlingMiddleware> _logger;
+
+        public ExceptionHandlingMiddleware(RequestDelegate next, ILogger<ExceptionHandlingMiddleware> logger)
+        {
+            _next = next;
+            _logger = logger;
+        }
+
+        public async Task Invoke(HttpContext context)
+        {
+            try
+            {
+                await _next(context); // 繼續下個 middleware 或 controller
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex, "Unhandled exception occurred.");
+
+                context.Response.ContentType = "application/json";
+                context.Response.StatusCode = (int)HttpStatusCode.InternalServerError;
+
+                var response = new
+                {
+                    message = ex.Message,
+                    statusCode = context.Response.StatusCode
+                };
+
+                await context.Response.WriteAsync(JsonSerializer.Serialize(response));
+            }
+        }
+    }
+}
